agent/requirements.txt
langgraph
langchain
langchain-google-genai
langchain-mongodb
pymongo
voyageai
python-dotenv
fastapi
uvicorn
pydantic
